sql_agent_backend/requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
langgraph
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
